backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
anthropic
pydantic-ai[anthropic]
policyengine-uk-compiled>=0.6.1
policyengine_uk
pandas
httpx
supabase
stripe
